refactor: swaps the filenames of the deprecated and new user methods
