+++ app_user.go
@@ -1,401 +1,219 @@
 package passage
 
 import (
-	"context"
-	"net/http"
-	"strings"
+	"errors"
 )
 
-type PassageUser = User
-type CreateUserArgs = CreateUserBody
-type UpdateUserArgs = UpdateBody
-
-type user struct {
-	appID  string
-	client *ClientWithResponses
-}
-
-func newUser(appID string, client *ClientWithResponses) *user {
-	return &user{
-		appID:  appID,
-		client: client,
-	}
-}
-
-// Get retrieves a user's object using their user ID.
-func (u *user) Get(userID string) (*PassageUser, error) {
-	res, err := u.client.GetUserWithResponse(context.Background(), u.appID, userID)
+// GetUser gets a user using their userID
+// returns user on success, error on failure
+//
+// Deprecated: Use `Passage.User.Get` instead.
+func (a *App) GetUser(userID string) (*User, error) {
+	user, err := a.User.Get(userID)
 	if err != nil {
-		return nil, err
+		var passageError PassageError
+		if errors.As(err, &passageError) {
+			return user, Error{
+				ErrorText:  passageError.Message,
+				ErrorCode:  passageError.ErrorCode,
+				Message:    passageError.Message,
+				StatusCode: passageError.StatusCode,
+			}
+		}
 	}
 
-	if res.JSON200 != nil {
-		return &res.JSON200.User, nil
-	}
-
-	var errorCode string
-	var message string
-	switch {
-	case res.JSON401 != nil:
-		errorCode = string(res.JSON401.Code)
-		message = res.JSON401.Error
-	case res.JSON404 != nil:
-		errorCode = string(res.JSON404.Code)
-		message = res.JSON404.Error
-	case res.JSON500 != nil:
-		errorCode = string(res.JSON500.Code)
-		message = res.JSON500.Error
-	}
-
-	return nil, PassageError{
-		ErrorCode:  errorCode,
-		Message:    message,
-		StatusCode: res.StatusCode(),
-	}
+	return user, err
 }
 
-// GetByIdentifier retrieves a user's object using their user identifier.
-func (u *user) GetByIdentifier(identifier string) (*PassageUser, error) {
-	limit := 1
-	lowerIdentifier := strings.ToLower(identifier)
-	res, err := u.client.ListPaginatedUsersWithResponse(
-		context.Background(),
-		u.appID,
-		&ListPaginatedUsersParams{
-			Limit:      &limit,
-			Identifier: &lowerIdentifier,
-		},
-	)
-
+// GetUserByIdentifier gets a user using their identifier
+// returns user on success, error on failure
+//
+// Deprecated: Use `Passage.User.GetByIdentifier` instead.
+func (a *App) GetUserByIdentifier(identifier string) (*User, error) {
+	user, err := a.User.GetByIdentifier(identifier)
 	if err != nil {
-		return nil, err
+		var passageError PassageError
+		if errors.As(err, &passageError) {
+			return user, Error{
+				ErrorText:  passageError.Message,
+				ErrorCode:  passageError.ErrorCode,
+				Message:    passageError.Message,
+				StatusCode: passageError.StatusCode,
+			}
+		}
 	}
 
-	if res.JSON200 != nil {
-		users := res.JSON200.Users
+	return user, err
+}
 
-		if len(users) == 0 {
-			return nil, PassageError{
-				ErrorCode:  "user_not_found",
-				Message:    "Could not find user with that identifier.",
-				StatusCode: http.StatusNotFound,
+// ActivateUser activates a user using their userID
+// returns user on success, error on failure
+//
+// Deprecated: Use `Passage.User.Activate` instead.
+func (a *App) ActivateUser(userID string) (*User, error) {
+	user, err := a.User.Activate(userID)
+	if err != nil {
+		var passageError PassageError
+		if errors.As(err, &passageError) {
+			return user, Error{
+				ErrorText:  passageError.Message,
+				ErrorCode:  passageError.ErrorCode,
+				Message:    passageError.Message,
+				StatusCode: passageError.StatusCode,
+			}
+		}
+	}
+
+	return user, err
+}
+
+// DeactivateUser deactivates a user using their userID
+// returns user on success, error on failure
+//
+// Deprecated: Use `Passage.User.Deactivate` instead.
+func (a *App) DeactivateUser(userID string) (*User, error) {
+	user, err := a.User.Deactivate(userID)
+	if err != nil {
+		var passageError PassageError
+		if errors.As(err, &passageError) {
+			return user, Error{
+				ErrorText:  passageError.Message,
+				ErrorCode:  passageError.ErrorCode,
+				Message:    passageError.Message,
+				StatusCode: passageError.StatusCode,
+			}
+		}
+	}
+
+	return user, err
+}
+
+// UpdateUser receives an UpdateBody struct, updating the corresponding user's attribute(s)
+// returns user on success, error on failure
+//
+// Deprecated: Use `Passage.User.Update` instead.
+func (a *App) UpdateUser(userID string, updateBody UpdateBody) (*User, error) {
+	user, err := a.User.Update(userID, updateBody)
+	if err != nil {
+		var passageError PassageError
+		if errors.As(err, &passageError) {
+			return user, Error{
+				ErrorText:  passageError.Message,
+				ErrorCode:  passageError.ErrorCode,
+				Message:    passageError.Message,
+				StatusCode: passageError.StatusCode,
+			}
+		}
+	}
+
+	return user, err
+}
+
+// DeleteUser receives a userID (string), and deletes the corresponding user
+// returns true on success, false and error on failure (bool, err)
+//
+// Deprecated: Use `Passage.User.Delete` instead.
+func (a *App) DeleteUser(userID string) (bool, error) {
+	if err := a.User.Delete(userID); err != nil {
+		var passageError PassageError
+		if errors.As(err, &passageError) {
+			return false, Error{
+				ErrorText:  passageError.Message,
+				ErrorCode:  passageError.ErrorCode,
+				Message:    passageError.Message,
+				StatusCode: passageError.StatusCode,
 			}
 		}
 
-		return u.Get(users[0].ID)
+		return false, err
 	}
 
-	var errorCode string
-	var message string
-	switch {
-	case res.JSON401 != nil:
-		errorCode = string(res.JSON401.Code)
-		message = res.JSON401.Error
-	case res.JSON404 != nil:
-		errorCode = string(res.JSON404.Code)
-		message = res.JSON404.Error
-	case res.JSON500 != nil:
-		errorCode = string(res.JSON500.Code)
-		message = res.JSON500.Error
-	}
-
-	return nil, PassageError{
-		ErrorCode:  errorCode,
-		Message:    message,
-		StatusCode: res.StatusCode(),
-	}
+	return true, nil
 }
 
-// Activate activates a user using their user ID.
-func (u *user) Activate(userID string) (*PassageUser, error) {
-	res, err := u.client.ActivateUserWithResponse(context.Background(), u.appID, userID)
+// CreateUser receives a CreateUserBody struct, creating a user with provided values
+// returns user on success, error on failure
+//
+// Deprecated: Use `Passage.User.Create` instead.
+func (a *App) CreateUser(createUserBody CreateUserBody) (*User, error) {
+	user, err := a.User.Create(createUserBody)
 	if err != nil {
-		return nil, err
+		var passageError PassageError
+		if errors.As(err, &passageError) {
+			return user, Error{
+				ErrorText:  passageError.Message,
+				ErrorCode:  passageError.ErrorCode,
+				Message:    passageError.Message,
+				StatusCode: passageError.StatusCode,
+			}
+		}
 	}
 
-	if res.JSON200 != nil {
-		return &res.JSON200.User, nil
-	}
-
-	var errorCode string
-	var message string
-	switch {
-	case res.JSON401 != nil:
-		errorCode = string(res.JSON401.Code)
-		message = res.JSON401.Error
-	case res.JSON404 != nil:
-		errorCode = string(res.JSON404.Code)
-		message = res.JSON404.Error
-	case res.JSON500 != nil:
-		errorCode = string(res.JSON500.Code)
-		message = res.JSON500.Error
-	}
-
-	return nil, PassageError{
-		ErrorCode:  errorCode,
-		Message:    message,
-		StatusCode: res.StatusCode(),
-	}
+	return user, err
 }
 
-// Deactivate deactivates a user using their user ID.
-func (u *user) Deactivate(userID string) (*PassageUser, error) {
-	res, err := u.client.DeactivateUserWithResponse(context.Background(), u.appID, userID)
+// ListUserDevices lists a user's devices
+// returns a list of devices on success, error on failure
+//
+// Deprecated: Use `Passage.User.ListDevices` instead.
+func (a *App) ListUserDevices(userID string) ([]WebAuthnDevices, error) {
+	devices, err := a.User.ListDevices(userID)
 	if err != nil {
-		return nil, err
+		var passageError PassageError
+		if errors.As(err, &passageError) {
+			return devices, Error{
+				ErrorText:  passageError.Message,
+				ErrorCode:  passageError.ErrorCode,
+				Message:    passageError.Message,
+				StatusCode: passageError.StatusCode,
+			}
+		}
 	}
 
-	if res.JSON200 != nil {
-		return &res.JSON200.User, nil
-	}
-
-	var errorCode string
-	var message string
-	switch {
-	case res.JSON401 != nil:
-		errorCode = string(res.JSON401.Code)
-		message = res.JSON401.Error
-	case res.JSON404 != nil:
-		errorCode = string(res.JSON404.Code)
-		message = res.JSON404.Error
-	case res.JSON500 != nil:
-		errorCode = string(res.JSON500.Code)
-		message = res.JSON500.Error
-	}
-
-	return nil, PassageError{
-		ErrorCode:  errorCode,
-		Message:    message,
-		StatusCode: res.StatusCode(),
-	}
+	return devices, err
 }
 
-// Update updates a user.
-func (u *user) Update(userID string, options UpdateUserArgs) (*PassageUser, error) {
-	res, err := u.client.UpdateUserWithResponse(context.Background(), u.appID, userID, options)
-	if err != nil {
-		return nil, err
+// RevokeUserDevice gets a user using their userID
+// returns a true success, error on failure
+//
+// Deprecated: Use `Passage.User.RevokeDevice` instead.
+func (a *App) RevokeUserDevice(userID, deviceID string) (bool, error) {
+	if err := a.User.RevokeDevice(userID, deviceID); err != nil {
+		var passageError PassageError
+		if errors.As(err, &passageError) {
+			return false, Error{
+				ErrorText:  passageError.Message,
+				ErrorCode:  passageError.ErrorCode,
+				Message:    passageError.Message,
+				StatusCode: passageError.StatusCode,
+			}
+		}
+
+		return false, err
 	}
 
-	if res.JSON200 != nil {
-		return &res.JSON200.User, nil
-	}
-
-	var errorCode string
-	var message string
-	switch {
-	case res.JSON400 != nil:
-		errorCode = string(res.JSON400.Code)
-		message = res.JSON400.Error
-	case res.JSON401 != nil:
-		errorCode = string(res.JSON401.Code)
-		message = res.JSON401.Error
-	case res.JSON404 != nil:
-		errorCode = string(res.JSON404.Code)
-		message = res.JSON404.Error
-	case res.JSON500 != nil:
-		errorCode = string(res.JSON500.Code)
-		message = res.JSON500.Error
-	}
-
-	return nil, PassageError{
-		ErrorCode:  errorCode,
-		Message:    message,
-		StatusCode: res.StatusCode(),
-	}
+	return true, nil
 }
 
-// Create creates a user.
-func (u *user) Create(args CreateUserArgs) (*PassageUser, error) {
-	res, err := u.client.CreateUserWithResponse(context.Background(), u.appID, args)
-	if err != nil {
-		return nil, err
+// Signout revokes a users refresh tokens
+// returns true on success, error on failure
+//
+// Deprecated: Use `Passage.User.RevokeRefreshTokens` instead.
+func (a *App) SignOut(userID string) (bool, error) {
+	if err := a.User.RevokeRefreshTokens(userID); err != nil {
+
+		var passageError PassageError
+		if errors.As(err, &passageError) {
+			return false, Error{
+				ErrorText:  passageError.Message,
+				ErrorCode:  passageError.ErrorCode,
+				Message:    passageError.Message,
+				StatusCode: passageError.StatusCode,
+			}
+		}
+
+		return false, err
 	}
 
-	if res.JSON201 != nil {
-		return &res.JSON201.User, nil
-	}
-
-	var errorText string
-	var errorCode string
-	switch {
-	case res.JSON400 != nil:
-		errorText = res.JSON400.Error
-		errorCode = string(res.JSON400.Code)
-	case res.JSON401 != nil:
-		errorText = res.JSON401.Error
-		errorCode = string(res.JSON401.Code)
-	case res.JSON404 != nil:
-		errorText = res.JSON404.Error
-		errorCode = string(res.JSON404.Code)
-	case res.JSON500 != nil:
-		errorText = res.JSON500.Error
-		errorCode = string(res.JSON500.Code)
-	}
-
-	return nil, Error{
-		Message:    errorText,
-		StatusCode: res.StatusCode(),
-		StatusText: res.Status(),
-		ErrorText:  errorText,
-		ErrorCode:  errorCode,
-	}
+	return true, nil
 }
-
-// Delete deletes a user using their user ID.
-func (u *user) Delete(userID string) error {
-	res, err := u.client.DeleteUserWithResponse(context.Background(), u.appID, userID)
-	if err != nil {
-		return err
-	}
-
-	if res.StatusCode() >= 200 && res.StatusCode() < 300 {
-		return nil
-	}
-
-	var errorCode string
-	var message string
-	switch {
-	case res.JSON401 != nil:
-		errorCode = string(res.JSON401.Code)
-		message = res.JSON401.Error
-	case res.JSON404 != nil:
-		errorCode = string(res.JSON404.Code)
-		message = res.JSON404.Error
-	case res.JSON500 != nil:
-		errorCode = string(res.JSON500.Code)
-		message = res.JSON500.Error
-	}
-
-	return PassageError{
-		ErrorCode:  errorCode,
-		Message:    message,
-		StatusCode: res.StatusCode(),
-	}
-}
-
-// ListDevices retrieves a user's webauthn devices using their user ID.
-func (u *user) ListDevices(userID string) ([]WebAuthnDevices, error) {
-	res, err := u.client.ListUserDevicesWithResponse(context.Background(), u.appID, userID)
-	if err != nil {
-		return nil, err
-	}
-
-	if res.JSON200 != nil {
-		return res.JSON200.Devices, nil
-	}
-
-	var errorCode string
-	var message string
-	switch {
-	case res.JSON401 != nil:
-		errorCode = string(res.JSON401.Code)
-		message = res.JSON401.Error
-	case res.JSON404 != nil:
-		errorCode = string(res.JSON404.Code)
-		message = res.JSON404.Error
-	case res.JSON500 != nil:
-		errorCode = string(res.JSON500.Code)
-		message = res.JSON500.Error
-	}
-
-	return nil, PassageError{
-		ErrorCode:  errorCode,
-		Message:    message,
-		StatusCode: res.StatusCode(),
-	}
-}
-
-// RevokeDevice revokes user's webauthn device using their user ID and the device ID.
-func (u *user) RevokeDevice(userID, deviceID string) error {
-	res, err := u.client.DeleteUserDevicesWithResponse(context.Background(), u.appID, userID, deviceID)
-	if err != nil {
-		return err
-	}
-
-	if res.StatusCode() >= 200 && res.StatusCode() < 300 {
-		return nil
-	}
-
-	var errorCode string
-	var message string
-	switch {
-	case res.JSON401 != nil:
-		errorCode = string(res.JSON401.Code)
-		message = res.JSON401.Error
-	case res.JSON404 != nil:
-		errorCode = string(res.JSON404.Code)
-		message = res.JSON404.Error
-	case res.JSON500 != nil:
-		errorCode = string(res.JSON500.Code)
-		message = res.JSON500.Error
-	}
-
-	return PassageError{
-		ErrorCode:  errorCode,
-		Message:    message,
-		StatusCode: res.StatusCode(),
-	}
-}
-
-// RevokeRefreshTokens revokes all of a user's Refresh Tokens using their User ID.
-func (u *user) RevokeRefreshTokens(userID string) error {
-	res, err := u.client.RevokeUserRefreshTokensWithResponse(context.Background(), u.appID, userID)
-	if err != nil {
-		return err
-	}
-
-	if res.StatusCode() >= 200 && res.StatusCode() < 300 {
-		return nil
-	}
-
-	var errorCode string
-	var message string
-	switch {
-	case res.JSON401 != nil:
-		errorCode = string(res.JSON401.Code)
-		message = res.JSON401.Error
-	case res.JSON404 != nil:
-		errorCode = string(res.JSON404.Code)
-		message = res.JSON404.Error
-	case res.JSON500 != nil:
-		errorCode = string(res.JSON500.Code)
-		message = res.JSON500.Error
-	}
-
-	return PassageError{
-		ErrorCode:  errorCode,
-		Message:    message,
-		StatusCode: res.StatusCode(),
-	}
-}
-
-// type userResponse struct {
-// 	JSON200 *UserResponse
-// 	JSON400 *N400Error
-// 	JSON401 *N401Error
-// 	JSON404 *N404Error
-// 	JSON500 *N500Error
-// }
-
-// func getPassageError[T userResponse](response T) PassageError {
-// 	var errorCode string
-// 	var message string
-// 	switch {
-// 	case response.JSON401 != nil:
-// 		errorCode = string(response.JSON401.Code)
-// 		message = response.JSON401.Error
-// 	case response.JSON404 != nil:
-// 		errorCode = string(response.JSON404.Code)
-// 		message = response.JSON404.Error
-// 	case response.JSON500 != nil:
-// 		errorCode = string(response.JSON500.Code)
-// 		message = response.JSON500.Error
-// 	}
-
-// 	return PassageError{
-// 		ErrorCode:  errorCode,
-// 		Message:    message,
-// 		StatusCode: response.StatusCode(),
-// 	}
-// }
